+++ cx/utilities.go
@@ -7,7 +7,6 @@ import (
 	"github.com/skycoin/cx/cx/globals"
 	"github.com/skycoin/cx/cx/helper"
 	"github.com/skycoin/cx/cx/mem"
-	"github.com/skycoin/cx/cx/tostring"
 )
 
 // Debug ...
@@ -22,96 +21,6 @@ func ExprOpName(expr *ast.CXExpression) string {
 	}
 	return expr.Operator.Name
 
-}
-
-func stackValueHeader(fileName string, fileLine int) string {
-	return fmt.Sprintf("%s:%d", fileName, fileLine)
-}
-
-// PrintStack ...
-func (cxprogram *ast.CXProgram) PrintStack() {
-	fmt.Println()
-	fmt.Println("===Callstack===")
-
-	// we're going backwards in the stack
-	fp := cxprogram.StackPointer
-
-	for c := cxprogram.CallCounter; c >= 0; c-- {
-		op := cxprogram.CallStack[c].Operator
-		fp -= op.Size
-
-		var dupNames []string
-
-		fmt.Printf(">>> %s()\n", op.Name)
-
-		for _, inp := range op.Inputs {
-			fmt.Println("ProgramInput")
-			fmt.Printf("\t%s : %s() : %s\n", stackValueHeader(inp.FileName, inp.FileLine), op.Name, tostring.GetPrintableValue(fp, inp))
-
-			dupNames = append(dupNames, inp.Package.Name+inp.Name)
-		}
-
-		for _, out := range op.Outputs {
-			fmt.Println("ProgramOutput")
-			fmt.Printf("\t%s : %s() : %s\n", stackValueHeader(out.FileName, out.FileLine), op.Name, tostring.GetPrintableValue(fp, out))
-
-			dupNames = append(dupNames, out.Package.Name+out.Name)
-		}
-
-		// fmt.Println("Expressions")
-		exprs := ""
-		for _, expr := range op.Expressions {
-			for _, inp := range expr.Inputs {
-				if inp.Name == "" || expr.Operator == nil {
-					continue
-				}
-				var dup bool
-				for _, name := range dupNames {
-					if name == inp.Package.Name+inp.Name {
-						dup = true
-						break
-					}
-				}
-				if dup {
-					continue
-				}
-
-				// fmt.Println("\t", inp.Name, "\t", ":", "\t", GetPrintableValue(fp, inp))
-				// exprs += fmt.Sprintln("\t", stackValueHeader(inp.FileName, inp.FileLine), "\t", ":", "\t", GetPrintableValue(fp, inp))
-
-				exprs += fmt.Sprintf("\t%s : %s() : %s\n", stackValueHeader(inp.FileName, inp.FileLine), ExprOpName(expr), tostring.GetPrintableValue(fp, inp))
-
-				dupNames = append(dupNames, inp.Package.Name+inp.Name)
-			}
-
-			for _, out := range expr.Outputs {
-				if out.Name == "" || expr.Operator == nil {
-					continue
-				}
-				var dup bool
-				for _, name := range dupNames {
-					if name == out.Package.Name+out.Name {
-						dup = true
-						break
-					}
-				}
-				if dup {
-					continue
-				}
-
-				// fmt.Println("\t", out.Name, "\t", ":", "\t", GetPrintableValue(fp, out))
-				// exprs += fmt.Sprintln("\t", stackValueHeader(out.FileName, out.FileLine), ":", GetPrintableValue(fp, out))
-
-				exprs += fmt.Sprintf("\t%s : %s() : %s\n", stackValueHeader(out.FileName, out.FileLine), ExprOpName(expr), tostring.GetPrintableValue(fp, out))
-
-				dupNames = append(dupNames, out.Package.Name+out.Name)
-			}
-		}
-
-		if len(exprs) > 0 {
-			fmt.Println("Expressions\n", exprs)
-		}
-	}
 }
 
 // IsTempVar ...
